# parser: Use an operator precedence table