@andrewrk 2021-04-25
> It may also be possible to speed this up by putting more useful information
> into the tokens during tokenization, to avoid the extra branch on the token in
> operInfo.

That's unlikely to happen. We have a really compact 5 bytes per token right now, and doing so was a measurable performance improvement. So far all the big perf wins have been by making representation of various pieces of heap data smaller.

The code simplification in this patch is nice. Can you share any performance impact stats?