@ifreund 2021-05-13
I've run some basic benchmarking formatting the std with perf and it looks like this PR is a very minor performance improvement if anything, certainly not a regression:
Before:
```
$ perf stat -e cache-misses -e instructions -e cycles ./build/zig fmt lib/std

 Performance counter stats for './build/zig fmt lib/std':

           787,097      cache-misses:u                                              
     4,988,785,455      instructions:u            #    2.89  insn per cycle         
     1,725,526,234      cycles:u                                                    

       0.415256074 seconds time elapsed

       0.402567000 seconds user
       0.011987000 seconds sys
```
After:
```
$ perf stat -e cache-misses -e instructions -e cycles ./build/zig fmt lib/std

 Performance counter stats for './build/zig fmt lib/std':

           818,776      cache-misses:u                                              
     4,800,522,512      instructions:u            #    2.91  insn per cycle         
     1,648,634,860      cycles:u                                                    

       0.399708854 seconds time elapsed

       0.378598000 seconds user
       0.019974000 seconds sys
```